+++ packages/at_client/CHANGELOG.md
@@ -1,6 +1,5 @@
 ## 3.0.50
-- feat: changes for at_chops uptake
-- chore: upgrade at_persistence_spec, at_persistence_secondary_server, at_commons version
+- feat: Introduce commit log compaction to keep size of commit log thin
 ## 3.0.49
 - fix: Enable AtKey.namespace overrides the namespace in AtClientPreference in AtClient delete method
 - fix: Fixed a bug where initial notifications fails to decrypt - invalid pad block issue